+++ R/fips.R
@@ -1,21 +1,15 @@
-#' FIPS URL
-#' 
-#' This file is an ASCII "dump" in pipe ("|") delimited form, of a "master" 
-#' shapefile maintained by AGMG for the maintenance of County, Public Forecast 
-#' Zones, CWA boundaries, and Time zones. Each record represents a single 
-#' polygon within the master file. Multiple polygons may exist for the same 
-#' zone/county that contain the same attributes, but the lat and lon fields 
-#' will be different as they are the centroid of the polygon. No attempt has 
-#' been made to remove these seemingly duplicate entries from this text file. 
-#' In addition, the coastal and offshore marine zone shapefiles are also 
-#' "dumped" to this file. 
-#'
+#' @title fips 
+#' @description FIPS URL
+#' @details This file is an ASCII "dump" in pipe ("|") delimited form, of a "master" 
+#'   shapefile maintained by AGMG for the maintenance of County, Public Forecast 
+#'   Zones, CWA boundaries, and Time zones. Each record represents a single 
+#'   polygon within the master file. Multiple polygons may exist for the same 
+#'   zone/county that contain the same attributes, but the lat and lon fields 
+#'   will be different as they are the centroid of the polygon. No attempt has 
+#'   been made to remove these seemingly duplicate entries from this text file. 
+#'   In addition, the coastal and offshore marine zone shapefiles are also 
+#'   "dumped" to this file. 
 #' @return url to fips dataset
-#' @export
-#' @examples
-#' 
-#' url <- fips()
-#' 
 fips <- function() {
     url <- "http://www.nws.noaa.gov/geodata/catalog/wsom/data/bp10nv15.dbx"
 }
@@ -25,65 +19,54 @@ fips <- function() {
 #' Helps set the names of Fips data table
 #'
 #' @return a character vector, length 10, of names to assign to fips_tz_dt()
-#' @export
 #' @examples
-#' \dontrun{
-#' names(fips_dt) <- fips_names()
-#' }
-#' 
+#'   \dontrun{
+#'     names(fips_dt) <- fips_names()
+#'   }
 fips_names <- function() {
     c("STATE", "ZONE", "CWA", "NAME", "STATE_ZONE", "COUNTYNAME", "FIPS", 
       "TIME_ZONE", "FE_AREA", "LAT", "LON")
 }
 
-#' Required fields for datetime objects
-#'
+#' @title datetime_fields 
+#' @description Required fields for datetime objects
 #' @return a character vector of required date/time variables
-#' @export
 #' @examples
-#' \dontrun{
-#' if(all(names(DT)) %in% datetime_fields())...
-#' 
-#' names(DT) <- datetime_fields()
-#' }
+#'   \dontrun{
+#'     if(all(names(DT)) %in% datetime_fields())...
+#'     names(DT) <- datetime_fields()
+#'   }
 datetime_fields <- function() {
     c("EVENT_ID", "YEAR", "MONTH_NAME", "BEGIN_DAY", "BEGIN_DATE_TIME", 
       "END_DAY", "END_DATE_TIME")
 }
 
-#' Required fields for timezone objects
-#'
+#' @title timezone_fields
+#' @description Required fields for timezone objects
 #' @return character vector of required timezone variables
-#' @export
 #' @examples
-#' \dontrun{
-#' if(all(names(DT)) %in% timezone_fields())...
-#' 
-#' names(DT) <- timezone_fields()
-#' }
+#'  \dontrun{
+#'     if(all(names(DT)) %in% timezone_fields())...
+#'     names(DT) <- timezone_fields()
+#'   }
 timezone_fields <- function() {
     c("EVENT_ID", "CZ_TIMEZONE")
 }
 
-#' Add Date/Time variables
-#' 
-#' Expects a data table with the variables EVENT_ID, TZ, YEAR, MONTH_NAME, 
-#'      BEGIN_DAY, BEGIN_DATE_TIME, END_DAY, and END_DATE_TIME. All variables 
-#'      except TZ come in the get_data() call requesting the \emph{details} 
-#'      dataset.
-#' 
-#' Two new variables are created and added to the data table: BEGIN_DATE and 
-#'      END_DATE. These dates are formatted like %Y-%m-%d %h:%m:%s %z. Then 
-#'      each variable has time readjusted to UTC with lubridate::with_tz(). 
-#'      The new variables are then added to LUDT. 
-#' 
+#' @title add_datetime 
+#' @description Add Date/Time variables
+#' @details Expects a data table with the variables EVENT_ID, TZ, YEAR, MONTH_NAME, 
+#'    BEGIN_DAY, BEGIN_DATE_TIME, END_DAY, and END_DATE_TIME. All variables 
+#'    except TZ come in the get_data() call requesting the \emph{details} 
+#'    dataset.
+#'    
+#'    Two new variables are created and added to the data table: BEGIN_DATE and 
+#'    END_DATE. These dates are formatted like %Y-%m-%d %h:%m:%s %z. Then 
+#'    each variable has time readjusted to UTC with lubridate::with_tz(). 
+#'    The new variables are then added to LUDT. 
 #' @return LUDT with datetime variables formatted to long-string format. 
-#' 
 #' @param DT data table consisting of EVENT_ID, TZ, YEAR, MONTH_NAME, 
 #'      BEGIN_DAY, BEGIN_DATE_TIME, END_DAY, END_DATE_TIME
-#'
-#' @export
-#' 
 add_datetime <- function(DT = NULL) {
     if(!is.data.table(DT)) stop("No data table.")
     if(!all(names(DT) %in% c("EVENT_ID", "TZ", "YEAR", "MONTH_NAME", 
@@ -126,10 +109,9 @@ add_datetime <- function(DT = NULL) {
 #'
 #' @param DT EVENT_ID and CZ_TIMEZONE field only
 #' @return Creates individual data tables for each unique timezone
-#' @export
 #' @examples
 #' \dontrun{
-#' tz_to_dt(DT[, .(EVENT_ID, CZ_TIMEZONE)])
+#'   tz_to_dt(DT[, .(EVENT_ID, CZ_TIMEZONE)])
 #' }
 tz_to_dt <- function(DT = NULL) {
     
@@ -152,18 +134,14 @@ tz_to_dt <- function(DT = NULL) {
 
 }
 
-#' Return FIPS dataset
-#'
-#' See \url{http://www.nws.noaa.gov/geodata/catalog/wsom/html/cntyzone.htm}
-#'
+#' @title fips_dt
+#' @description Return FIPS dataset
+#' @details See \url{http://www.nws.noaa.gov/geodata/catalog/wsom/html/cntyzone.htm}
 #' @return dataset of FIPS codes.
-#' @export
-#'
 #' @examples
-#' \dontrun{
-#' FIPS <- fips_dt()
-#' }
-#' 
+#'   \dontrun{
+#'     FIPS <- fips_dt()
+#'   }
 fips_dt <- function() {
     url <- fips()
     fips_dt <- as.data.table(readr::read_delim(url, delim = "|", 
@@ -193,14 +171,11 @@ fips_dt <- function() {
 #' }
 #' 
 #' @return a character vector to map fips_tz_dt$TIME_ZONE to the true 
-#' timezone
-#' 
-#' @export
-#'
+#'   timezone
 #' @examples
-#' \dontrun{
-#' tz_abbr <- fips_tz_abbr()
-#' }
+#'   \dontrun{
+#'     tz_abbr <- fips_tz_abbr()
+#'   }
 fips_tz_abbr <- function() {
     list("A" = "America/Anchorage", 
          "C" = "America/Chicago", 
@@ -213,17 +188,14 @@ fips_tz_abbr <- function() {
          "V" = "America/Halifax")
 }
 
-#' Add FIPS codes to LUDT
-#'
+#' @title add_fips
+#' @description Add FIPS codes to LUDT
 #' @param DT data table consisting of EVENT_ID, CZ_FIPS, STATE_FIPS
-#'
 #' @return adds FIPS codes to lookup data table associated w/ EVENT_ID
-#' @export
-#'
 #' @examples
-#' \dontrun{
-#' add_fips(DT[, .(EVENT_ID, CZ_FIPS, STATE_FIPS)])
-#' }
+#'   \dontrun{
+#'     add_fips(DT[, .(EVENT_ID, CZ_FIPS, STATE_FIPS)])
+#'   }
 add_fips <- function(DT = NULL) {
     if(!is.data.table(DT)) stop("No data table.")
     if(!all(names(DT) %in% c("EVENT_ID", "CZ_FIPS", "STATE_FIPS")))
@@ -241,18 +213,14 @@ create_LUDT <- function() {
     assign("LUDT", data.table("EVENT_ID" = as.numeric()), envir = .GlobalEnv)
 }
 
-#' Add time zones to lookup table
-#'
+#' @title add_tz 
+#' @description Add time zones to lookup table
 #' @param DT data table consisting of EVENT_ID, TZ
-#'
-#' @export
-#' 
 add_tz <- function(DT = NULL) {
-    if(!is.data.table(DT)) stop("No data table.")
-    if(!all(names(DT) %in% c("EVENT_ID", "TZ")))
-        stop("Expecting a data table with EVENT_ID, TZ")
-    if(!exists("LUDT")) create_LUDT()
-    LUDT <- merge(LUDT, DT, by = "EVENT_ID", all = TRUE)
-    LUDT
+  if(!is.data.table(DT)) stop("No data table.")
+  if(!all(names(DT) %in% c("EVENT_ID", "TZ")))
+    stop("Expecting a data table with EVENT_ID, TZ")
+  if(!exists("LUDT")) create_LUDT()
+  LUDT <- merge(LUDT, DT, by = "EVENT_ID", all = TRUE)
+  LUDT
 }
-