+++ CHANGELOG.md
@@ -9,6 +9,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
   - R/details.R - store code for processing details dataset
+  - R/fatalities.R - code for processing fatalities dataset
 
 ### Changed
   - NA